Rough-draft JWT Support
The following code is a quick swipe at what first-class support for
OAuth2 Resource Server flows might look like in Spring Security.
I'm starting only with JWT support while we flesh out the details.

Of particular debate is that Spring Security tends to push
HTTP requests into a two-step process of Authentication and then
Authorization. In the OAuth2 Resource Server use case, we really
only care about Authorization.

Still, the draft here works with Spring Security in the pattern
that it expects. One nuance under consideration is exception
handling.

Also, under investigation is how to improve on the SpEL support
of the legacy spring-security-oauth2, which relies on a sub-class
of DefaultMethodSecurityExpressionHandler. This is a bit
inflexible for users since we are spending their inheritance for
them in order to provide that functionality.

Note that the classes that I think are close are marked with a
@since, etc., but classes that I think might entirely disappear
under team scrutiny are just commented with some of my thoughts as to
where we might take things.