Moving Auth0 Integration to Sample
We want to pick more than one JWT library to integrate with the API
to make sure that it is abstracted appropriately. To do that, I'm
moving the Auth0 integration into the sample that was written for
it.

The plan is to add more samples to further stress test the API.