Reclassify to Verifiers and Formats
First, it appears to me that OAuth2 Access Tokens can be expressed in
a number of formats. From the JWT spec:

JSON Web Token (JWT) is a compact claims format

And in the case of an opaque token, the format is "opaque" to us
until, perhaps, it is processed.

For now, the DSL supports the configuration of JWT and Opaque as
formats supported by this resource server instance.

Roughly speaking, each format configured implies an additional
AuthenticationProvider.

In configuring each format, the DSL supports custom ways to process
tokens of that format.

Second, verification strategies stand independent of the token's
original format. Once any claims have actually been made, then they
can be verified.

At the very least, a token's claims to its own validity are
confirmed. For example, a token's iat and exp. This is not
configurable.

Additionally, there are two additional standard verification steps,
decrypting the claims and verifying the attached signature.

The DSL, then, supports configuration of these two verification
steps, though they will not result in instances of
OAuth2AccessTokenVerifier.

The DSL also allows for the specification of custom verification on
top of basic verification.

Issue: spring-projects/spring-security#5226