Initial Support for Single-Key Configuration
Adding another implementation of JwtDecoder for Nimbus, though I
imagine that we'd merge the new NimbusJwtDecoder somehow with
NimbusJwtDecoderJwkSupport, though precisely how is unclear.

Accompanied with this commit is a working sample that configures a
resource server with a single key. To facilitate this, I've moved
JwsBuilder and JwkSetBuilder out of src/test/java and into
src/main/java, though I imagine that I will replace this with some
other abstraction, perhaps something like what Rob has mentioned in

Finally, this commit includes some support for building a JWT that
aligns itself with SS's assumptions about type and defaults for each
JWT claim, see JwtBuilder. If we keep this, I would actually imagine
it being a subclass inside Jwt, and thus folks could use Jwt.Builder
when it suited their needs (as it did mine in NimbusJwtDecoder).

Issue: #20