Initial insufficient_scope support
Support for this feature is not as clean as I'd like it to be at this
point, though it's not yet clear to me how else I would do it without
engaging in a risky refactor of the access denial pipeline in Spring
Security Core.

As it stands, each hasXXXScope method will throw an
OAuth2AccessDeniedException instead of returning false. This is
configurable as well. There is now a BearerTokenAccessDeniedHandler
which will look for this exception and respond accordingly.

Also, I've combined the bearer token error support from the entry
point and the access denied handler into BearerTokenErrorHandler.

Issue: spring-projects/spring-security#5227