ResourceServerConfigurer on Auth0 Sample
Reconfigured the auth0 sample to use the Resource Server Configurer.

Note that this uncovered a problem with BearerTokenErrorHandler where
is was calling sendError instead of setStatus. The contract of
sendError indicates that it converts the content type to text/html and
allows the container to send an html error response with the given
message.

Calling setStatus seems more apppropriate in our case.

Issue: spring-projects/spring-security#5125