Send invalid request with invalid bearer token
The issue recomends escaping the . that is inside the character class
in the regular expression; however, this appears to be unecessary. A
test was added to confirm.

However, the issue also pointed out that the token was simply being
ignored if it failed the regex. This commit changes that so that if
the token contains invalid characters, then an error is returned to
the user.

Fixes: gh-28