Simplify Resource Server Configurer
ResourceServerConfigurer no longer needs to be BeanFactoryAware,
though it does need an instance of SessionManagementConfigurer in
its constructor in order to configure it early enough in the building
process.

This is a clear departure from the other configurers, so I plan to do
more research to see if I can further simplify it.

Issue: gh-13