AuthoritiesExtractor
A need that has come up multiple times now is to extract the roles
(or scopes) from an OAuth2 token in some custom way. This isn't
authority mapping since we aren't taking a list of GrantedAuthorities
and converting them into another list.

So, this commit introduces the concept of authority extraction into
the resource server.

Issue: gh-37