Key Resolver
This commit introduces a generic way to resolve keys, which allows
for a wider array of strategies for identifying keys.

Implementers will be given the JWT header and can decide on which
key to provide accordingly.

Note that this could be a hook for folks who need more sophisticated
URL support than just configuring an endpoint.