Utilities Project
This is a project including support classes for the various sample
projects. It may be that these will be promoted and released elsewhere
within the Spring Security project.

You'll notice a few things in this project:

1. Helper classes for creating key pairs, and reading and writing them
to and from the filesystem
2. Command-line utilities for doing the above
3. ConversionService support for PublicKey and PrivateKey

This introduction of this project is tied directly to converting some
of the sample applications into ones that use the filesystem instead
of just runtime-generated keys.

Issue: gh-42