Evolve Custom Scope Attribute Support
User can now configure what will be the expected attribute name for
scopes on incoming JWTs.

If unspecified, the code will try and determine based on the heuristic
that if it finds an attribute of 'scp' in the token, then it will
assume scp.

The 'scp' attribute is also now moved into a constant, though it needs
a more permanent home. It's currently unclear where it will go since
JwtClaimNames is reserved for attributes in the JWT spec exclusively.

Fixes: gh-54