Rename OAuth2TokenVerifier
I realize that we are still pretty early and it is even still possible
that this class will be sanded away through the additional refactoring
rounds; however, I believe this class will stick around, and I do
agree with the team that this is a better name. I figure I'm willing
to risk it this time around in order to prevent a more impactful
refactor downstream.

Fixes: gh-56