Token Validators + Sample
Here is a sample usage of validators as well and some refactoring
of validator usage including the introduction of a delegating
validator.

Not sure if these will stay, but this commit also adds a few what may
be standard validators like JwtClaimValidator.

Also note that this feels awfully similar to some of the work that is
done over on the SpEL side of things. Perhaps this is an indication
that they should be sharing logic in some way.

Issue: gh-52, gh-50, gh-51