Auto-configuration Second Draft
Refactored to make the auto configuration mimic the structure of
OAuth2ClientAutoConfiguration and its reactive counterpart. Still
left is to consider whether or not the more extensive use of the DSL
is merited or if, like the other auto configurations, it would be
better to resolve to the underlying beans and expose them, e.g with
JwtDecoder.

Issue: gh-46