Refine Validator Configuration
Validators are now configured (via the DSL) one at a time by chaining
.validator calls instead of passing a list to .validators. This
provides more flexibility for disparate code locations to weigh in on
which validators are in play as well as being a bit more readable.

Also, I removed the ValidatorConfigurer class. In the end, I felt like
it was assuming too much about how users will in the end use the DSL.
Perhaps we may offer higher-level representations in the DSL for
issuer, audience, and the like in future releases.