Add Clock to Timestamp Validation
Now Clock can be used in testing environments instead of creating new
tokens with new timestamps each time.

Note that Nimbus itself performs timestamp validation as well, which
does not use a Clock, so this validation is bypassed in the local
key support Nimbus implementation. A similar change would be necessary
to provide this feature across the board.

Issue: gh-57