AuthoritiesExtractor -> AuthoritiesPopulator
The facility for extracting authorities now simply populates an
Authentication with authorities. This will enable users, by default,
to see scopes as granted authorities and it will allow them to perform
more sophisticated transformations.

Also as a result, the scope attribute configuration was pushed down to
the JwtAuthoritiesPopulator since the authentication provider no
longer needs it.

Fixes: gh-37